requirements.txt
aiohttp
asyncio
discord.py
pynacl
youtube-dl
python-dotenv